Simplify ConsumerInfo
This is a preparation for connecting the Consumer with Interpreter. The
new ConsumerInfo contains almost every information needed to be
serialized about a Consumer. The only missing part is the reference to
the transformation, but it has to be handled in a different way,
because it has to store the transformation identifier (module.trans),
not the concrete implementation (which is ConsumerFunction).