Adapt RsmClient into un-copyable type
Since RsmClient holds two separate MGFutures as members, objects of
this type can not be copied anymore. This change caused the project not
to compile, because RsmClient instances were copied in some of the
simulation tests. This commit fixes those tests, deletes the copy ctors
of RsmClient and defines the move assignement operator for MGFuture.