Fix failing unit tests
The issue was caused by the difference when trying to recover from
snapshots and WAL files. The parallelized version of index creation is
based on RecoveryInfo::vertex_batches which is always empty in the case
of recovering from WAL files. In that case the sequential index creation
will be used.