Fix work with RocksDB column families. Resolve stack-use-after-scope in storage. Disable GC for disk storage because we remove data after each tx.
It's prohibited to destroy handle to default CF of RocksDB if it was got by db_->DefaultColumnFamily(). In this case we see double free corruption. But we are required to destroy the handle if we open an existing database and got new handle from Open() function.
The reason is Open() creates another handle with reference to CF, so before closing database we need to delete all references which were created by Memgraph code.