Initial commit of GraphQL Support
Initial version of an e2e testing framework that can be used to run
graphql queries end to end with memgraph. For the 'create' graphql query
to work, the awesome memgraph function, randomUUID, had to be added.
Further more a temporary query module is loaded in the workload to
emulate the initial query to the database. This should be changed later.