Implement util.validate, make '_' a valid varname
The util.validate procedure could not be added as a simple query
procedure, because the procedure itself does not return anything, which
is something we do not support atm, and we do not have the ability to
evaluate expressions from query procedures. They it has been solved, is
that the during parsing, we check if the given CALL procedure does not
YIELD, and if the fully qualified name of the procedure matches
util.validate and based on this, we can instantiate an appropriate
cursur and put that into the pull-chain instead of the regular
CallProcedureCursor. On top of this '_' is sometimes generated as a
variable name we did not support. Support for that has been added as
well.