Fix concurrent query module racecondition
Concurrent access to the same query module had a racecondition on the
pointer that was used to handle the costum memory management. With this
commit, a mapping has been added that keeps information about what
thread used what pointer to handle the memory resources. Which should be
fine since the respected query executions are running on a dedicated
thread. Access to the mapping itself is threadsafe. A simple spinlock
has been implemented here, as we shouldn't include memgraph::utils from
this header and a traditional mutex might be overkill. A simple RAII
wrapper for the mapping container has been also added for simpler
client-side use.