Change caching mechanism.
Instead of trying to read into the index caches directly, read into the
main chache, and merge the index and main caches when appropriate. This
is needed because of several index modifying calls within one
transaction makes the indexing process convoluted, and should be
propertly addressed when the On-disk LRU cache will be properly
implemented.