Fix double send of deltas to GC
The mechanism for unique_constraints->Validate requires a commit
timestamp to be set on the accessor before we actually commit. But it is
possible that we have a violation and need to Abort.

In this case we have a commit_timestamp_ which FinalizeTransaction uses
as an indicator that it needs to send the deltas to
committed_transactions so that it can be GC'ed.

The fix, after Abort ensure we clear commit_timestamp_. In Abort we
already unlinked and sent to GC as well as MarkFinished.
FinalizeTransaction doesn't need to do anything.