Add initial commit for the edge indexing feature
Currently there are two different kinds of indices in the database
indices for label and indiced for label+property based lookups on
vertices. With this feature it should be possible to create and deal
with indices on edge types in the same way as with indices on vertices.
The already existing implementations of indices are good as a start for
indices for edge-types, however they are not mappable one to one. Issue
with edge-type based indices is that it is not always clearly defines
what en edge is. If the flag --storage-properties-enabled-on-edges is
False, then there is no singular datastrucrture that is the collection
of all the edges present in the database. The only way to get the edges
is to gather all the information runtime using the source and
destination vertices of the given edge, and compute everything we need
to be able to construct the EdgeAccessor objects. This might be
computationally to expensive, especially in the presence of supernodes.
If the datastructure that is the collection of all the edges is enabled
by setting the --storage-properties-enabled-on-edges to True, it is
still not trivial to get all the data that is needed to create the
EdgeAccessor objects. Solution for these issues should come in
subsequent commits.