Clean up on-disk stubs
Previously abstract base classes were used to implement the on-disk
related stubs as we do not plan on implementing this functionality for
now for the on-disk storage mode. With this commit the relevant virtual
functions have been made pure virtual and the stuv implementations have
been moved to dedicated subclasses.