Make indexing work while inserting new edges
The edge-type index creation itself is scanning through the skiplist and
building up the indices based on that, however it does not take into
account those edges that are being created after the index itself is
created. With this commit the index is updated upon adding edges to the
database if needed.