Add more tests and resolve transactional issues
The complexity brought by storage::View::NEW and OLD are making some
edgecases hard to deal with. Add neccessary metadata, needed by the edge
indexing process to work correctly in these edge-cases. Currently that
is an additional pointer to the Edge constructs in the Edges skiplist.
Add more comprehensive tests for the indexing and planning.