Add recovery capabilities
Add recovery capabilites so the edge indices are also loaded in the case
when snapshot creation is triggered or the recovery process is
initiated. The way of retrieveing edges had to be modifed. When
obtaining the edges from the index only the existence of the edge with
the given edge-type between two nodes was checked and returned. This was
buggy, because unlike labels on vertices, there can be more than one
edge-typed edges between two nodes, so additional logic had to be added
to fix this behavior.