Replace throwing with return values
In the case of ON_DISK_TRANSACTIONAL storage mode, the operations that
supposed to check if edge-type indexing is available usedto throw a not
yet implemented error. This is a bug, since the function that determines
if the edge-type indexing could be used is throwing, propagating the
error to the user, instead of returning false and let the planner pick
the non-indexing plan.