Bugfix: Replica now handles skipped transactions
On REPLICA the commit timestamp is calculated in an interesting way.
This is for reasons related to durability based on the timestamp. The
problem is that some timestamps can be skipped, and hence never used on
the replica side and hence never finalised. This would mean the commit
log would provide incorrect information to GC about what the oldest
active timestamp is. Causing GC not to clear away the Deltas being made.

Fix is to make all skipped timestamps to be marked as finalised.

Another memory related issue was the RPC session buffer retained the
size of its high watermark. Post Session::Execute we now limit the
buffer back to a max of 4MiB.