Execute state check on becoming leader (#2248)
When becoming the leader in case everything seems okay we don't do any
checks. But the coordinator can become leader and only do commits of
older logs afterwards (when we exit become leader callback), meaning we
could miss logs which put us in unhealthy state.

To fix that problem, we will alter become leader callback to always do
check of whole cluster and set it in healthy state in the beginning.