Improve fine grained access e2e tests (#2279)
This PR removes redundant tests in the fine_grained_access end-to-end
tests. Specifically, I eliminated tests that served the same purpose.
The original tests included cases for the MATCH, CREATE, and MERGE
clauses. Since the MERGE clause essentially performs a MATCH, and if no
match is found, it proceeds to CREATE, I believe it is unnecessary to
test MERGE separately in this context. Any issues with the MERGE
statement should be addressed in the dedicated tests for MERGE.
Additionally, added some small improvements to `common.py`.