Edge index planning in sync with variable start planner (#2293)
Edge index planning made robust with the variable start planner like the
node planning.
Before this PR, the plans could only be generated with scanning of
nodes, followed by expansion. Since we introduced edge indices based on
edge type and edge-property type, the planner needs to be adjusted so it
can generate a scanning of indexed edge triplets (from, edge, to) in
order to perform valid expansion.
If there is no edge index present, the plan will not be used, since we
don't have a data structure that will iterate over all the edges and
their triplets. That plan is invalid. All plans starting from an edge
index are valid.