Start Nemesis with 5' delay and run Jepsen tests separately (#2330)
Jepsen tests we run on diff are now executed separately to be able to
better control tests' input. test-all-individually task from run.sh
script is deleted since we now use the test function. Generator's code
is refactored to be able to better understand in what sequence are
generator's operations sequenced. Nemesis process now starts with 5'
delay so that the setup of the cluster can finish before any disruptions
start occurring.