Replication fixes (#2283)
Bugfix: Wrong recovery steps when snapshot needs to be used
Bugfix: Applying snapshot on REPLICA does not update timestamp, which
leads to older delta getting applied
Bugfix: Erasure of older WAL files could also remove the current WAL
(due to it also removing malformed files)
Stopped checking REPLICA for the exact timestamp, because snapshot
carries only the storage timestamp of the transaction used to generate
the snapshot. A REPLICA can appear to be ahead of MAIN because of this.
Since we protect REPLICAs from multiple MAINs, this shouldn't be a
problem.
Applying a snapshot resets the WAL sequence number as well. Everything
gets deleted before applying a snapshot, so the old seq number doesn't
make sense anymore.