RTLD_DEEPBIND removal (#2347)
In the past we had issues with RTLD_DEEPBIND.
Some issues made a query module work because of RTLD_DEEPBIND, others
would break because of RTLD_DEEPBIND.

Advice is to remove RTLD_DEEPBIND and make most global symbols hidden. 

My belief is that most of the issues we were having were solved by the
update to toolchain-v5 which used a newer libstdc++. I'm currently
unable to reproduce the issue, so removing RTLD_DEEPBIND is a first
step. Fixing symbol hiding will be done in a later PR.