Fix txn handling (#2360)
Nemesis could kill the thread between committing a transaction and
incrementing the atomic counter. This caused the issue because multiple
requests for creating same range of indices could be sent when the txn
got committed and we didn't register that change. The issue is solved by
finding the maximum current vertex id and creating nodes in the same
transaction.